% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{cb_authentication_vars}
\alias{cb_authentication_vars}
\title{TODO}
\usage{
cb_authentication_vars(package, which, hint = TRUE, ...)
}
\arguments{
\item{package}{TODO}

\item{which}{TODO}
}
\value{
TODO
}
\description{
TODO
}
