% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_prepare.R
\name{cb_climate_prepare_job}
\alias{cb_climate_prepare_job}
\title{TODO}
\usage{
cb_climate_prepare_job(product_id, ...)
}
\arguments{
\item{product_id}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
\examples{
if (interactive()) {
  cb_climate_prepare_job("reanalysis-era5-land")
}
}
