% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{cb_insert_authentication_vars}
\alias{cb_insert_authentication_vars}
\title{TODO}
\usage{
cb_insert_authentication_vars(...)
}
\arguments{
\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
